refactor: update alias command
